backend/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
cloud-init==24.3.1
cloudpathlib==0.20.0
colorama==0.4.4
command-not-found==0.3
confection==0.1.5
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cymem==2.0.8
dbus-python==1.2.18
deprecation==2.1.0
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
frozenlist==1.5.0
gotrue==2.11.0
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.20.2
httpx==0.27.2
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.7.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
langcodes==3.4.1
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.1.0
murmurhash==1.0.10
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==2.0.2
oauthlib==3.2.0
openai==1.54.4
outcome==1.3.0.post0
packaging==24.1
postgrest==0.18.0
preshed==3.0.9
propcache==0.2.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==5.1.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
realtime==2.0.6
requests==2.32.3
rich==13.9.4
SecretStorage==3.3.1
service-identity==18.1.0
shellingham==1.5.4
simplejson==3.17.6
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
storage3==0.9.0
StrEnum==0.4.15
supabase==2.10.0
supafunc==0.7.0
systemd-python==234
thinc==8.3.2
tqdm==4.66.6
trio==0.27.0
trio-websocket==0.11.1
Twisted==22.1.0
typer==0.12.5
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
yarl==1.17.1
zipp==1.0.0
zope.interface==5.4.0
